tools/database-migrator/requirements.txt
fire==0.7.0
importlib_metadata==8.5.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
sqlparse==0.5.3
tabulate==0.9.0
termcolor==2.5.0
yoyo-migrations==9.0.0
zipp==3.21.0
